requirements.txt
git+https://github.com/LiBa001/disputils
git+https://github.com/Pycord-Development/pycord
aiohttp
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
disputils==0.2.0
gcloud==0.18.3
giphy-client==1.0.0
googleapis-common-protos==1.53.0
httplib2==0.20.2
idna==3.3
imageio==2.12.0
imgurpython==1.1.7
jws==0.1.3
multidict==5.2.0
oauth2client==4.1.3
praw==7.5.0
prawcore==2.3.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.11.0
pymongo==3.12.1
pyparsing==3.0.4
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python-jwt==2.0.1
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.2
update-checker==0.18.0
urbandict==0.6.1
urllib3==1.26.7
websocket-client==1.2.1
yarl==1.7.2
akinator.py[async]
pynacl 
youtube-dl
readabledelta
petpetgif